Use uint64_t instead of unsigned long long

+++ dummy/energymon-dummy.h
@@ -11,18 +11,19 @@
 extern "C" {
 #endif
 
+#include <inttypes.h>
 #include <stddef.h>
 #include "energymon.h"
 
 int energymon_init_dummy(energymon* em);
 
-unsigned long long energymon_read_total_dummy(const energymon* em);
+uint64_t energymon_read_total_dummy(const energymon* em);
 
 int energymon_finish_dummy(energymon* em);
 
 char* energymon_get_source_dummy(char* buffer, size_t n);
 
-unsigned long long energymon_get_interval_dummy(const energymon* em);
+uint64_t energymon_get_interval_dummy(const energymon* em);
 
 int energymon_get_dummy(energymon* impl);
