+++ test/interval_test.c
@@ -1,4 +1,5 @@
 #include <assert.h>
+#include <inttypes.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <unistd.h>
@@ -7,7 +8,7 @@
 int main(int argc, char** argv) {
   char source[100];
   long usec;
-  unsigned long long start, end, uj;
+  uint64_t start, end, uj;
   if (argc < 2) {
     fprintf(stderr, "Usage: %s <idle_usec>\n", argv[0]);
     return 1;
@@ -27,7 +28,7 @@ int main(int argc, char** argv) {
   end = impl.fread(&impl);
   uj = end - start;
 
-  printf("Total energy: %lld uJ\n", uj);
+  printf("Total energy: %"PRIu64" uJ\n", uj);
   printf("Average power: %f W\n", (double) uj / (double) usec);
   assert(impl.ffinish(&impl) == 0);
   printf("Finished reading from %s\n", source);
